+++ core/src/main/java/org/nervos/appchain/tx/ManagedTransaction.java
@@ -38,8 +38,8 @@ public abstract class ManagedTransaction {
 
     // adapt to cita
     protected TransactionReceipt sendAdaptToCita(
-            String to, String data, BigInteger quota, BigInteger nonce,
-            BigInteger validUntilBlock, BigInteger version , int chainId, String value)
+            String to, String data, long quota, BigInteger nonce,
+            long validUntilBlock, int version , int chainId, String value)
             throws IOException, TransactionException {
         return transactionManager.executeTransaction(
                 to, data, quota, nonce, validUntilBlock, version, chainId, value);
